Task/recipient event emitter (#289)
